requirements.txt
# CATENA Journal Scraper Dependencies

# Web scraping - Playwright version
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Web scraping - Selenium version (alternative, better anti-detection)
undetected-chromedriver>=3.5.0
selenium>=4.15.0

# Date parsing
python-dateutil>=2.8.0

# Data export
openpyxl>=3.1.0

# HTTP requests (backup)
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
tqdm>=4.66.0

# Retry logic
tenacity>=8.2.0
